backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-multipart
pydantic
Pillow
numpy
torch
torchvision
